third_party/WebKit/LayoutTests/fast/encoding/api/streaming-decode-expected.txt

Test streaming decoding using the Encoding API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


string = '\x00123ABCabc\x80\xFF\u0100\u1000\uFFFD\uD800\uDC00\uDBFF\uDFFF'

encoded = new TextEncoder('utf-8').encode(string)
out = ''
decoder = new TextDecoder('utf-8')
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([49]), {stream: true})
out += decoder.decode(new Uint8Array([50]), {stream: true})
out += decoder.decode(new Uint8Array([51]), {stream: true})
out += decoder.decode(new Uint8Array([65]), {stream: true})
out += decoder.decode(new Uint8Array([66]), {stream: true})
out += decoder.decode(new Uint8Array([67]), {stream: true})
out += decoder.decode(new Uint8Array([97]), {stream: true})
out += decoder.decode(new Uint8Array([98]), {stream: true})
out += decoder.decode(new Uint8Array([99]), {stream: true})
out += decoder.decode(new Uint8Array([194]), {stream: true})
out += decoder.decode(new Uint8Array([128]), {stream: true})
out += decoder.decode(new Uint8Array([195]), {stream: true})
out += decoder.decode(new Uint8Array([191]), {stream: true})
out += decoder.decode(new Uint8Array([196]), {stream: true})
out += decoder.decode(new Uint8Array([128]), {stream: true})
out += decoder.decode(new Uint8Array([225]), {stream: true})
out += decoder.decode(new Uint8Array([128]), {stream: true})
out += decoder.decode(new Uint8Array([128]), {stream: true})
out += decoder.decode(new Uint8Array([239]), {stream: true})
out += decoder.decode(new Uint8Array([191]), {stream: true})
out += decoder.decode(new Uint8Array([189]), {stream: true})
out += decoder.decode(new Uint8Array([240]), {stream: true})
out += decoder.decode(new Uint8Array([144]), {stream: true})
out += decoder.decode(new Uint8Array([128]), {stream: true})
out += decoder.decode(new Uint8Array([128]), {stream: true})
out += decoder.decode(new Uint8Array([244]), {stream: true})
out += decoder.decode(new Uint8Array([143]), {stream: true})
out += decoder.decode(new Uint8Array([191]), {stream: true})
out += decoder.decode(new Uint8Array([191]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-8')
out += decoder.decode(new Uint8Array([0,49]), {stream: true})
out += decoder.decode(new Uint8Array([50,51]), {stream: true})
out += decoder.decode(new Uint8Array([65,66]), {stream: true})
out += decoder.decode(new Uint8Array([67,97]), {stream: true})
out += decoder.decode(new Uint8Array([98,99]), {stream: true})
out += decoder.decode(new Uint8Array([194,128]), {stream: true})
out += decoder.decode(new Uint8Array([195,191]), {stream: true})
out += decoder.decode(new Uint8Array([196,128]), {stream: true})
out += decoder.decode(new Uint8Array([225,128]), {stream: true})
out += decoder.decode(new Uint8Array([128,239]), {stream: true})
out += decoder.decode(new Uint8Array([191,189]), {stream: true})
out += decoder.decode(new Uint8Array([240,144]), {stream: true})
out += decoder.decode(new Uint8Array([128,128]), {stream: true})
out += decoder.decode(new Uint8Array([244,143]), {stream: true})
out += decoder.decode(new Uint8Array([191,191]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-8')
out += decoder.decode(new Uint8Array([0,49,50]), {stream: true})
out += decoder.decode(new Uint8Array([51,65,66]), {stream: true})
out += decoder.decode(new Uint8Array([67,97,98]), {stream: true})
out += decoder.decode(new Uint8Array([99,194,128]), {stream: true})
out += decoder.decode(new Uint8Array([195,191,196]), {stream: true})
out += decoder.decode(new Uint8Array([128,225,128]), {stream: true})
out += decoder.decode(new Uint8Array([128,239,191]), {stream: true})
out += decoder.decode(new Uint8Array([189,240,144]), {stream: true})
out += decoder.decode(new Uint8Array([128,128,244]), {stream: true})
out += decoder.decode(new Uint8Array([143,191,191]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-8')
out += decoder.decode(new Uint8Array([0,49,50,51]), {stream: true})
out += decoder.decode(new Uint8Array([65,66,67,97]), {stream: true})
out += decoder.decode(new Uint8Array([98,99,194,128]), {stream: true})
out += decoder.decode(new Uint8Array([195,191,196,128]), {stream: true})
out += decoder.decode(new Uint8Array([225,128,128,239]), {stream: true})
out += decoder.decode(new Uint8Array([191,189,240,144]), {stream: true})
out += decoder.decode(new Uint8Array([128,128,244,143]), {stream: true})
out += decoder.decode(new Uint8Array([191,191]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-8')
out += decoder.decode(new Uint8Array([0,49,50,51,65]), {stream: true})
out += decoder.decode(new Uint8Array([66,67,97,98,99]), {stream: true})
out += decoder.decode(new Uint8Array([194,128,195,191,196]), {stream: true})
out += decoder.decode(new Uint8Array([128,225,128,128,239]), {stream: true})
out += decoder.decode(new Uint8Array([191,189,240,144,128]), {stream: true})
out += decoder.decode(new Uint8Array([128,244,143,191,191]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"

encoded = new TextEncoder('utf-16le').encode(string)
out = ''
decoder = new TextDecoder('utf-16le')
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([49]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([50]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([51]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([65]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([66]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([67]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([97]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([98]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([99]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([128]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([255]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([1]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([16]), {stream: true})
out += decoder.decode(new Uint8Array([253]), {stream: true})
out += decoder.decode(new Uint8Array([255]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([216]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([220]), {stream: true})
out += decoder.decode(new Uint8Array([255]), {stream: true})
out += decoder.decode(new Uint8Array([219]), {stream: true})
out += decoder.decode(new Uint8Array([255]), {stream: true})
out += decoder.decode(new Uint8Array([223]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-16le')
out += decoder.decode(new Uint8Array([0,0]), {stream: true})
out += decoder.decode(new Uint8Array([49,0]), {stream: true})
out += decoder.decode(new Uint8Array([50,0]), {stream: true})
out += decoder.decode(new Uint8Array([51,0]), {stream: true})
out += decoder.decode(new Uint8Array([65,0]), {stream: true})
out += decoder.decode(new Uint8Array([66,0]), {stream: true})
out += decoder.decode(new Uint8Array([67,0]), {stream: true})
out += decoder.decode(new Uint8Array([97,0]), {stream: true})
out += decoder.decode(new Uint8Array([98,0]), {stream: true})
out += decoder.decode(new Uint8Array([99,0]), {stream: true})
out += decoder.decode(new Uint8Array([128,0]), {stream: true})
out += decoder.decode(new Uint8Array([255,0]), {stream: true})
out += decoder.decode(new Uint8Array([0,1]), {stream: true})
out += decoder.decode(new Uint8Array([0,16]), {stream: true})
out += decoder.decode(new Uint8Array([253,255]), {stream: true})
out += decoder.decode(new Uint8Array([0,216]), {stream: true})
out += decoder.decode(new Uint8Array([0,220]), {stream: true})
out += decoder.decode(new Uint8Array([255,219]), {stream: true})
out += decoder.decode(new Uint8Array([255,223]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-16le')
out += decoder.decode(new Uint8Array([0,0,49]), {stream: true})
out += decoder.decode(new Uint8Array([0,50,0]), {stream: true})
out += decoder.decode(new Uint8Array([51,0,65]), {stream: true})
out += decoder.decode(new Uint8Array([0,66,0]), {stream: true})
out += decoder.decode(new Uint8Array([67,0,97]), {stream: true})
out += decoder.decode(new Uint8Array([0,98,0]), {stream: true})
out += decoder.decode(new Uint8Array([99,0,128]), {stream: true})
out += decoder.decode(new Uint8Array([0,255,0]), {stream: true})
out += decoder.decode(new Uint8Array([0,1,0]), {stream: true})
out += decoder.decode(new Uint8Array([16,253,255]), {stream: true})
out += decoder.decode(new Uint8Array([0,216,0]), {stream: true})
out += decoder.decode(new Uint8Array([220,255,219]), {stream: true})
out += decoder.decode(new Uint8Array([255,223]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-16le')
out += decoder.decode(new Uint8Array([0,0,49,0]), {stream: true})
out += decoder.decode(new Uint8Array([50,0,51,0]), {stream: true})
out += decoder.decode(new Uint8Array([65,0,66,0]), {stream: true})
out += decoder.decode(new Uint8Array([67,0,97,0]), {stream: true})
out += decoder.decode(new Uint8Array([98,0,99,0]), {stream: true})
out += decoder.decode(new Uint8Array([128,0,255,0]), {stream: true})
out += decoder.decode(new Uint8Array([0,1,0,16]), {stream: true})
out += decoder.decode(new Uint8Array([253,255,0,216]), {stream: true})
out += decoder.decode(new Uint8Array([0,220,255,219]), {stream: true})
out += decoder.decode(new Uint8Array([255,223]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-16le')
out += decoder.decode(new Uint8Array([0,0,49,0,50]), {stream: true})
out += decoder.decode(new Uint8Array([0,51,0,65,0]), {stream: true})
out += decoder.decode(new Uint8Array([66,0,67,0,97]), {stream: true})
out += decoder.decode(new Uint8Array([0,98,0,99,0]), {stream: true})
out += decoder.decode(new Uint8Array([128,0,255,0,0]), {stream: true})
out += decoder.decode(new Uint8Array([1,0,16,253,255]), {stream: true})
out += decoder.decode(new Uint8Array([0,216,0,220,255]), {stream: true})
out += decoder.decode(new Uint8Array([219,255,223]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"

encoded = new TextEncoder('utf-16be').encode(string)
out = ''
decoder = new TextDecoder('utf-16be')
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([49]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([50]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([51]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([65]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([66]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([67]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([97]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([98]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([99]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([128]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([255]), {stream: true})
out += decoder.decode(new Uint8Array([1]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([16]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([255]), {stream: true})
out += decoder.decode(new Uint8Array([253]), {stream: true})
out += decoder.decode(new Uint8Array([216]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([220]), {stream: true})
out += decoder.decode(new Uint8Array([0]), {stream: true})
out += decoder.decode(new Uint8Array([219]), {stream: true})
out += decoder.decode(new Uint8Array([255]), {stream: true})
out += decoder.decode(new Uint8Array([223]), {stream: true})
out += decoder.decode(new Uint8Array([255]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-16be')
out += decoder.decode(new Uint8Array([0,0]), {stream: true})
out += decoder.decode(new Uint8Array([0,49]), {stream: true})
out += decoder.decode(new Uint8Array([0,50]), {stream: true})
out += decoder.decode(new Uint8Array([0,51]), {stream: true})
out += decoder.decode(new Uint8Array([0,65]), {stream: true})
out += decoder.decode(new Uint8Array([0,66]), {stream: true})
out += decoder.decode(new Uint8Array([0,67]), {stream: true})
out += decoder.decode(new Uint8Array([0,97]), {stream: true})
out += decoder.decode(new Uint8Array([0,98]), {stream: true})
out += decoder.decode(new Uint8Array([0,99]), {stream: true})
out += decoder.decode(new Uint8Array([0,128]), {stream: true})
out += decoder.decode(new Uint8Array([0,255]), {stream: true})
out += decoder.decode(new Uint8Array([1,0]), {stream: true})
out += decoder.decode(new Uint8Array([16,0]), {stream: true})
out += decoder.decode(new Uint8Array([255,253]), {stream: true})
out += decoder.decode(new Uint8Array([216,0]), {stream: true})
out += decoder.decode(new Uint8Array([220,0]), {stream: true})
out += decoder.decode(new Uint8Array([219,255]), {stream: true})
out += decoder.decode(new Uint8Array([223,255]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-16be')
out += decoder.decode(new Uint8Array([0,0,0]), {stream: true})
out += decoder.decode(new Uint8Array([49,0,50]), {stream: true})
out += decoder.decode(new Uint8Array([0,51,0]), {stream: true})
out += decoder.decode(new Uint8Array([65,0,66]), {stream: true})
out += decoder.decode(new Uint8Array([0,67,0]), {stream: true})
out += decoder.decode(new Uint8Array([97,0,98]), {stream: true})
out += decoder.decode(new Uint8Array([0,99,0]), {stream: true})
out += decoder.decode(new Uint8Array([128,0,255]), {stream: true})
out += decoder.decode(new Uint8Array([1,0,16]), {stream: true})
out += decoder.decode(new Uint8Array([0,255,253]), {stream: true})
out += decoder.decode(new Uint8Array([216,0,220]), {stream: true})
out += decoder.decode(new Uint8Array([0,219,255]), {stream: true})
out += decoder.decode(new Uint8Array([223,255]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-16be')
out += decoder.decode(new Uint8Array([0,0,0,49]), {stream: true})
out += decoder.decode(new Uint8Array([0,50,0,51]), {stream: true})
out += decoder.decode(new Uint8Array([0,65,0,66]), {stream: true})
out += decoder.decode(new Uint8Array([0,67,0,97]), {stream: true})
out += decoder.decode(new Uint8Array([0,98,0,99]), {stream: true})
out += decoder.decode(new Uint8Array([0,128,0,255]), {stream: true})
out += decoder.decode(new Uint8Array([1,0,16,0]), {stream: true})
out += decoder.decode(new Uint8Array([255,253,216,0]), {stream: true})
out += decoder.decode(new Uint8Array([220,0,219,255]), {stream: true})
out += decoder.decode(new Uint8Array([223,255]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
out = ''
decoder = new TextDecoder('utf-16be')
out += decoder.decode(new Uint8Array([0,0,0,49,0]), {stream: true})
out += decoder.decode(new Uint8Array([50,0,51,0,65]), {stream: true})
out += decoder.decode(new Uint8Array([0,66,0,67,0]), {stream: true})
out += decoder.decode(new Uint8Array([97,0,98,0,99]), {stream: true})
out += decoder.decode(new Uint8Array([0,128,0,255,1]), {stream: true})
out += decoder.decode(new Uint8Array([0,16,0,255,253]), {stream: true})
out += decoder.decode(new Uint8Array([216,0,220,0,219]), {stream: true})
out += decoder.decode(new Uint8Array([255,223,255]), {stream: true})
out += decoder.decode()
PASS out is "\u0000123ABCabc每锟金靠"
PASS successfullyParsed is true

TEST COMPLETE
